separation of components and project folder architecture
