+++ src/App.js
@@ -1,0 +1,11 @@
+import React from 'react';
+import TweetList from './components/Tweet/TweetList'
+
+
+function App() {
+  return (
+    <TweetList/>
+  )
+}
+
+export default App;
