+++ src/components/Avatar.css
@@ -1,0 +1,6 @@
+.avatar {
+  width: 48px;
+  height: 48px;
+  border-radius: 50%;
+  margin-right: 10px; 
+}
