+++ src/components/Buttons/Buttons.css
@@ -1,0 +1,21 @@
+.buttons {
+  margin-top: 20px;
+  margin-left: 2px; 
+  font-size: 1.4em;
+}
+
+.buttons span {
+  display: inline-flex;
+  align-items: center; 
+  margin-right: 40px;
+  cursor: pointer;
+}
+
+.buttons span span {
+  margin-right: 5px;
+}
+
+.buttons label {
+  font-size: 16px; 
+  font-weight: bold;
+}
