+++ src/components/Buttons/OptionsBtn.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import './Buttons.css';
+
+const OptionsBtn = () => (
+  <span className="material-icons grey-color-1">more_horiz</span>
+);
+
+export default OptionsBtn;
