+++ src/components/Buttons/ReplyBtn.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import './Buttons.css';
+
+const ReplyBtn = () => (
+  <span className="material-icons grey-color-1">reply</span>
+);
+
+export default ReplyBtn;
