+++ src/components/Buttons/VerifiedIcon.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const VerifiedIcon = () => (
+  <span className="material-icons primary-color">verified</span>
+);
+
+export default VerifiedIcon;
