+++ src/components/Tweet/Message.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+function Message({ text }) {
+  return (
+    <div style={{marginTop: "5px"}}>
+      {text}
+    </div>
+  );
+}
+
+export default Message;
