+++ src/components/Tweet/Tweet.css
@@ -1,0 +1,11 @@
+.tweet {
+  display: flex;
+  width: 564px;
+  min-height: 68px;
+  margin: 50px auto;
+  padding: 10px;
+  font-family: "Helvetica", arial, sans-serif; 
+  font-size: 16px;
+  line-height: 18px;
+  border: 1px solid #ccc;
+}
