+++ src/components/Tweet/TweetList.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import Tweet from './Tweet';
+import tweetData from '../../data/TweetData';
+
+function TweetList() {
+  return (
+    <Tweet tweet={tweetData} />
+  )
+}
+
+export default TweetList;
+
