+++ src/data/TweetData.js
@@ -1,0 +1,29 @@
+const tweetData = {
+  message: "Wealth consists not in having great possessions, but in having few wants...",
+  avatar: "epictetus-pic.jpg",
+  author: {
+    handle: "stoic-philosopher",
+    name: "Epictetus", 
+    verified: true,
+  },
+  likes: 7, 
+  retweets: 12,
+  timestamp: "2020-08-10 21:24:37" 
+};
+
+export default tweetData; 
+
+
+
+// const tweetData = {
+//   message: "Wealth consists not in having great possessions, but in having few wants...",
+//   avatar: "epictetus-pic.jpg",
+//   author: {
+//     handle: "stoic-philosopher",
+//     name: "Epictetus", 
+//     verified: true,
+//   },
+//   likes: 7, 
+//   retweets: 12,
+//   timestamp: "2020-08-10 21:24:37" 
+// };
