+++ src/index.css
@@ -1,79 +1,4 @@
-/* body {
-  margin: 0 auto;
-} */
-
-.tweet {
-  display: flex;
-  width: 564px;
-  min-height: 68px;
-  margin: 50px auto;
-  padding: 10px;
-  font-family: "Helvetica", arial, sans-serif; 
-  font-size: 16px;
-  line-height: 18px;
-  border: 1px solid #ccc;
-}
-
-.avatar {
-  width: 48px;
-  height: 48px;
-  border-radius: 50%;
-  margin-right: 10px; 
-}
-
-
-.author-container {
-  display: flex;
-  align-items: center;
-} 
-
-.author {
-  display: flex;
-  align-items: center;
-}
-
-.name {
-  font-weight: bold;
-  margin-right: 2px;
-  /* margin-bottom: .5em; */
-  /* margin-right: .3em; */
-}
-
-.handle {
-  color: #8899a6;
-  margin-left: 5px;
-}
-
-.time {
-  padding-left: .3em;
-  color: #8899a6;
-}
-
-.message {
-  margin-top: 5px;
-}
-
-.buttons {
-  margin-top: 20px;
-  margin-left: 2px; 
-  font-size: 1.4em;
-  color: #aab8c2;
-}
-
-.buttons span {
-  display: inline-flex;
-  align-items: center; 
-  margin-right: 40px;
-  cursor: pointer;
-}
-
-.buttons span span {
-  margin-right: 5px;
-}
-
-.buttons label {
-  font-size: 16px; 
-  font-weight: bold;
-}
-
+/* COLORS */
 .primary-color {color: #46b8f8}
+.grey-color-1 {color: #aab8c2;}
+.grey-color-2 {color: #8899a6;}
